Fully working password reset

+++ src/main/java/com/teashop/teashop_backend/model/user/PasswordResetTokenRepository.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 
 @Repository
 public interface PasswordResetTokenRepository extends JpaRepository<PasswordResetToken, Long> {
-    Optional<PasswordResetToken> findById(int id);
+    Optional<PasswordResetToken> findById(Long id);
     Optional<PasswordResetToken> findByToken(String token);
     Optional<PasswordResetToken> findByUser(User user);
     void deleteByUser(User user);
